Initial commit of minimal working functionality
A container launched from this built image will:

* Run the Mirth Connect Administrator, listening for HTTP/HTTPS traffic on
  the default ports 80 and 443, respectively
* Run a NGiNX server on port 3000 which proxies requests to Mirth Connect,
  forwaring to port 443 if it receives `X-Forwarded-Proto: https`

In this way, the image supports being run directly, or behind a sequence
of reverse proxies (as is the case with Aptible Enclave Endpoints). Note
that since Mirth Connect is listening internally on ports 80/443, the user-
facing reverse proxy must also listen on ports 80/443, since Mirth Connect
will insist on rewriting all URLs to include the ports it believes itself
to be listening on. :sideeye:

By default, this image will use an in-container Derby database for storage
of message data and configuration. It also supports using a PostgreSQL
database, if a valid database URL is present in `$DATABASE_URL`.